+++ packages/components/src/registration/SignInForm.tsx
@@ -7,15 +7,11 @@ import { FormProvider, useForm } from "react-hook-form";
 import { z } from "zod";
 
 import { trpc } from "@good-dog/trpc/client";
+import { zSignInValues } from "@good-dog/trpc/schema";
 import { Checkbox } from "@good-dog/ui/checkbox";
 
 import GenericRegistrationForm from "./GenericRegistrationForm";
 import RegistrationInput from "./inputs/RegistrationInput";
-
-const zSignInValues = z.object({
-  email: z.string().email(),
-  password: z.string(),
-});
 
 type FormValues = z.infer<typeof zSignInValues>;
