+++ packages/trpc/src/schema/README.md
@@ -1,0 +1,5 @@
+# Schema
+
+This directory contains the schema definitions for the endpoints and frontend forms. These schemas should be shared between the backend and frontend to simplify validation
+
+Try to organize the schemas in a way that makes sense, and then export them in the [index.ts](./index.ts) file. This will allow you to import them easily in other files.
