+++ packages/trpc/src/types.ts
@@ -1,10 +1,23 @@
 import type { TRPCClientErrorLike } from "@trpc/client";
 import type { inferProcedureOutput } from "@trpc/server";
 
-import type { AppRouter } from "./internal/router";
+import type { Role } from "@good-dog/db";
 
-export type { UserWithSession } from "./internal/common-types";
+import type { AppRouter } from "./internal/router";
 
 export type TRPCErrorLike = TRPCClientErrorLike<AppRouter>;
 export type GetProcedureOutput<T extends keyof AppRouter> =
   inferProcedureOutput<AppRouter[T]>;
+
+export interface UserWithSession {
+  userId: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+  phoneNumber: string;
+  role: Role;
+  session: {
+    expiresAt: Date;
+    refreshRequired: boolean;
+  };
+}
